Extract React internals caches and nuke dead/WIP code
